Renamed stream_processors -> providers
Fixed mypy error
